requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
dnspython==2.4.2
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.0
object-detection==0.1
packaging==23.1
pymongo==4.5.0
pytz==2023.3.post1
PyYAML==6.0.1
slim==0.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
